@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 
 class InfluencingFeature {
-  InfluencingFeature({@required this.name, this.isPeriodic});
+  InfluencingFeature({required this.name, this.isPeriodic});
 
   String name;
-  bool isPeriodic = false;
+  bool? isPeriodic = false;
 }
