@@ -20,8 +20,8 @@ import 'package:psychodynamics/back/entities/influencer.dart';
 
 class MoodRecord {
   MoodRecord({this.comment,
-    @required this.influencers,
-    @required this.emotions, this.dt}) {
+    required this.influencers,
+    required this.emotions, this.dt}) {
     dt = DateTime.now();
   }
 
@@ -34,24 +34,24 @@ class MoodRecord {
 
   // describes what user generally feels
   // -100 <= generalValue <= 100
-  int generalValue;
+  int? generalValue;
 
   // TODO add this feature later
   // int socialInteractions;
-  String comment;
+  String? comment;
   List<Influencer> influencers;
   List<Emotion> emotions;
-  DateTime dt /*= DateTime.now()*/;
+  DateTime? dt /*= DateTime.now()*/;
 
   String get textForTile => general;
 
   String get general {
     if (generalValue == 0) return 'ok';
-    if (generalValue > 0) {
-      if (generalValue<50) return 'good';
+    if (generalValue! > 0) {
+      if (generalValue!<50) return 'good';
       return 'perfect';
     } else {
-      if (generalValue>-50) return 'bad';
+      if (generalValue!>-50) return 'bad';
       return 'really bad';
     }
   }
