@@ -4,7 +4,7 @@ import 'package:psychodynamics/front/ui_helper.dart';
 import 'package:psychodynamics/front/widgets/app_bar.dart';
 
 class HomePage extends StatefulWidget {
-  HomePage({Key key}):super(key: key);
+  HomePage({Key? key}) : super(key: key);
 
   static const routeName = '/home';
 
@@ -16,40 +16,46 @@ class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        automaticallyImplyLeading: false,
-        elevation: 0,
-        backgroundColor: Colors.transparent,
-        title: Text(
-          'Home',
-          style: TextStyle(
-            fontWeight: FontWeight.bold,
-            fontSize: 32,
+        appBar: AppBar(
+          automaticallyImplyLeading: false,
+          elevation: 0,
+          backgroundColor: Colors.transparent,
+          title: Text(
+            'Home',
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+              fontSize: 32,
+            ),
           ),
         ),
-      ),
-      backgroundColor: Theme.of(context).backgroundColor,
-      body: SafeArea(
-        child: SizedBox.expand(
-          child: Padding(
-            padding: EdgeInsets.symmetric(vertical: 10, horizontal: 20),
-            child: Column(
-              children: [
-                Container(
+        backgroundColor: Theme.of(context).backgroundColor,
+        body: SafeArea(
+          child: SizedBox.expand(
+            child: Padding(
+              padding: EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+              child: Column(
+                children: [
+                  Container(
                     decoration: BoxDecoration(
                       borderRadius: BorderRadius.circular(4),
                       color: primarySwatch,
                     ),
                     child: Padding(
                       padding: const EdgeInsets.all(10.0),
-                      child: Text('Welcome', style: TextStyle(fontWeight: FontWeight.w700, fontSize: 40, color: Colors.white),),
-                    )),
-              ],
+                      child: Text(
+                        'Welcome',
+                        style: TextStyle(
+                            fontWeight: FontWeight.w700,
+                            fontSize: 40,
+                            color: Colors.white),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
             ),
           ),
         ),
-      ),
-      bottomNavigationBar: CustomAppBar()
-    );
+        bottomNavigationBar: CustomAppBar());
   }
 }
